first

+++ controllers/adminController.js
@@ -10,11 +10,13 @@ const { validationResult } = require("express-validator");
 const validate = require("../middleware/custom_validator");
 const { sendResponse } = require("../util/common");
 const { name } = require("ejs");
+const log = require("../util/logFile");
 //const validate = require("../middleware/validation")
 const SECRET_KEY = "myapi";
 class admin {
     async addBook(req, res) {
         try {
+            log.createLogFile("/book/addBook", "Success")
             const validation = validationResult(req).array();
             if (validation.length > 0) {
                 return sendResponse(res, HTTP_STATUS.UNPROCESSABLE_ENTITY, "Failed to add the book", validation);
@@ -29,12 +31,14 @@ class admin {
             return sendResponse(res, HTTP_STATUS.OK, "Successfully stored data", products);
 
         } catch (error) {
+            log.createLogFile("/book/addBook", "Failure")
             console.log(error);
             return sendResponse(res, HTTP_STATUS.INTERNAL_SERVER_ERROR, "Internal Server Error!");
         }
     }
     async deleteBook(req, res) {
         try {
+            log.createLogFile("/book/deleteBook", "Success")
             const { id } = req.params;
             const findBook = await bookModel.findById({ _id: id });
             console.log(id);
@@ -45,22 +49,30 @@ class admin {
             return sendResponse(res, HTTP_STATUS.OK, "Book has been deleted successfully");
 
         } catch (error) {
+            log.createLogFile("/book/deleteBook", "Failure")
             console.log(error);
             return sendResponse(res, HTTP_STATUS.INTERNAL_SERVER_ERROR, "Internal Server Error!");
         }
     }
     async viewUserData(req, res) {
         try {
+            log.createLogFile("/book/viewUserData", "Success")
             const filter = {};
             const { id, name, email, order_number, page, limit } = req.query;
             let parsedpage = parseInt(page) || 1;
             let parsedlimit = parseInt(limit) || 10;
             const viewUsers = await userModel.find({}).select("-createdAt -updatedAt");
+            const accept = ["id", "name", "email", "order_number", "page", "limit"];
+            const wrongParam = Object.keys(req.query).filter((x) => !accept.includes(x));
+            console.log(`wrong ${wrongParam}`)
+            if (wrongParam.length > 0) {
+                return sendResponse(res, HTTP_STATUS.CONFLICT, "Request Invalid!");
+            }
             if (id) {
                 filter._id = id;
             }
             if (name) {
-                filter.name = name;
+                filter.name = { $regex: name, $options: "i" };
             }
             if (email) {
                 filter.email = email;
@@ -76,19 +88,27 @@ class admin {
                 return sendResponse(res, HTTP_STATUS.NOT_FOUND, "Unable to fetch data!")
             }
             if (findUsers.length == 0) {
-                return sendResponse(res, HTTP_STATUS.NO_CONTENT, "No user found!");
+                return sendResponse(res, HTTP_STATUS.NOT_FOUND, "No user found!");
             }
             return sendResponse(res, HTTP_STATUS.OK, "Successfully fetched data", findUsers);
         } catch (error) {
+            log.createLogFile("/book/viewUserData", "Failure")
             console.log(error);
             return sendResponse(res, HTTP_STATUS.INTERNAL_SERVER_ERROR, "Internal Server Error!");
         }
     }
     async deleteUser(req, res) {
         try {
+            log.createLogFile("/book/deleteUser", "Success")
             const { id } = req.query;
             const findUserinUser = await userModel.findById({ _id: id });
             const findUserinAuth = await authModel.findOne({ user: id });
+            const accept = ["id"];
+            const wrongParam = Object.keys(req.query).filter((x) => !accept.includes(x));
+            console.log(`wrong ${wrongParam}`)
+            if (wrongParam.length > 0) {
+                return sendResponse(res, HTTP_STATUS.CONFLICT, "Request Invalid!");
+            }
             if (findUserinUser && findUserinAuth) {
                 await userModel.deleteOne({ _id: id });
                 await authModel.deleteOne({ user: id });
@@ -97,6 +117,7 @@ class admin {
             return sendResponse(res, HTTP_STATUS.NOT_FOUND, "User not found!");
 
         } catch (error) {
+            log.createLogFile("/book/deleteUser", "Failure")
             console.log(error);
             return sendResponse(res, HTTP_STATUS.INTERNAL_SERVER_ERROR, "Internal Server Error!");
         }
@@ -104,10 +125,21 @@ class admin {
     async updateBook(req, res) {
 
         try {
-            const { id } = req.query
+            log.createLogFile("/book/updateBook", "Success")
+            const { _id } = req.query;
+            console.log(_id);
             const updateFields = {};
+            const { isbn, name, price, category, stock, author, publisher, discountPercentage, discountStart, discountEnd } = req.body;
+            const validation = validationResult(req).array();
+            if (validation.length > 0) {
+                return sendResponse(res, HTTP_STATUS.UNPROCESSABLE_ENTITY, "Failed to update the book", validation);
+            }
             if (isbn) {
                 updateFields.isbn = isbn;
+                const findIsbn = await bookModel.findOne({ isbn: isbn })
+                if (findIsbn && findIsbn._id != _id) {
+                    return sendResponse(res, HTTP_STATUS.CONFLICT, "ISBN already exist!");
+                }
             }
             if (name) {
                 updateFields.name = name;
@@ -127,33 +159,60 @@ class admin {
             if (publisher) {
                 updateFields.publisher = publisher;
             }
+            if (discountPercentage) {
+                updateFields.discountPercentage = discountPercentage;
+            }
+            if (discountStart) {
+                updateFields.discountStart = discountStart;
+            }
+            if (discountEnd) {
+                updateFields.discountEnd = discountEnd
+            }
             console.log(updateFields)
-            const updateResult = await bookModel.updateOne({ _id: id }, { $set: updateFields });
+            const updateResult = await bookModel.updateOne({ _id: _id }, { $set: updateFields });
             console.log(updateResult);
+            if (updateResult.modifiedCount === 0) {
+                return sendResponse(res, HTTP_STATUS.EXPECTATION_FAILED, "No document was updated");
+            }
             return sendResponse(res, HTTP_STATUS.OK, "Data updated successfully!", updateResult);
 
         } catch (error) {
+            log.createLogFile("/book/updateBook", "failure")
+            console.log(error)
             return sendResponse(res, HTTP_STATUS.INTERNAL_SERVER_ERROR, "Internal Server Error!");
         }
     }
     async updateUser(req, res) {
-
         try {
+            log.createLogFile("/book/updateUser", "Success")
             const { id } = req.query
             const updateFieldsAuth = {};
             const { name, email, phone, balance } = req.body;
+            const validation = validationResult(req).array();
+            if (validation.length > 0) {
+                return sendResponse(res, HTTP_STATUS.UNPROCESSABLE_ENTITY, "Failed to update the user", validation);
+            }
             if (name) {
                 updateFieldsAuth.name = name;
             }
             if (email) {
                 updateFieldsAuth.email = email;
+                const existEmail = await userModel.findOne({ email: email })
+                if (existEmail && existEmail._id != id) {
+                    return sendResponse(res, HTTP_STATUS.CONFLICT, "Email already registered");
+                }
             }
             if (phone) {
                 updateFieldsAuth.phone = phone;
+                const existPhone = await userModel.findOne({ phone: phone })
+                if (existPhone && existPhone._id != id) {
+                    return sendResponse(res, HTTP_STATUS.CONFLICT, "Phone Number already registered");
+                }
             }
             if (balance) {
                 updateFieldsAuth.balance = balance;
             }
+            console.log(`hooooooo ${updateFieldsAuth.name}`)
             if (Object.keys(updateFieldsAuth).length == 0) {
                 return sendResponse(res, HTTP_STATUS.NOT_ACCEPTABLE, "No data to update");
             }
@@ -165,14 +224,20 @@ class admin {
             return sendResponse(res, HTTP_STATUS.OK, "Data has been updated");
 
         } catch (error) {
+            log.createLogFile("/book/updateUser", "Failure")
             console.log(error);
             return sendResponse(res, HTTP_STATUS.INTERNAL_SERVER_ERROR, "Internal Server Error!");
         }
     }
     async addDiscount(req, res) {
         try {
+            log.createLogFile("/book/addDiscount", "Success")
             const { id, discountPercentage, discountStart, discountEnd } = req.body;
             const findProduct = await bookModel.findById({ _id: id });
+            const validation = validationResult(req).array();
+            if (validation.length > 0) {
+                return sendResponse(res, HTTP_STATUS.UNPROCESSABLE_ENTITY, "Failed to update the user", validation);
+            }
             if (!findProduct) {
                 return sendResponse(res, HTTP_STATUS.NOT_FOUND, "Product not found!");
             }
@@ -193,19 +258,27 @@ class admin {
                 await findProduct.save();
                 return sendResponse(res, HTTP_STATUS.OK, "Successfully stored data", result);
             }
+            return sendResponse(res, HTTP_STATUS.NOT_ACCEPTABLE, "Discount has been added already!");
 
         } catch (error) {
+            log.createLogFile("/book/addDiscount", "failure")
             console.log(`error now ${error}`)
             return sendResponse(res, HTTP_STATUS.INTERNAL_SERVER_ERROR, "Internal Server Error");
         }
     }
     async updateDiscount(req, res) {
         try {
-            const { id, discountPercentage, discountStart, discountEnd } = req.body;
+            log.createLogFile("/book/updateDiscount", "Success")
+            const { id } = req.query;
+            const { discountPercentage, discountStart, discountEnd } = req.body;
             const findDiscount = await discountModel.findOne({ productId: id });
             const findProduct = await bookModel.findById({ _id: id });
+            const validation = validationResult(req).array();
+            if (validation.length > 0) {
+                return sendResponse(res, HTTP_STATUS.UNPROCESSABLE_ENTITY, "Failed to update the user", validation);
+            }
             if (!findDiscount) {
-                return sendResponse(res, HTTP_STATUS.NOT_FOUND, "Not found!");
+                return sendResponse(res, HTTP_STATUS.NOT_FOUND, "Add discount first!");
             }
             const query = {};
             if (discountPercentage) {
@@ -226,12 +299,14 @@ class admin {
             return sendResponse(res, HTTP_STATUS.OK, "Discount updatd successfully!", updatedResult);
 
         } catch (error) {
+            log.createLogFile("/book/updateDiscount", "failure")
             console.log(error);
             return sendResponse(res, HTTP_STATUS.INTERNAL_SERVER_ERROR, "Internal Server Error!");
         }
     }
     async viewAllTransaction(req, res) {
         try {
+            log.createLogFile("/book/viewAllTransaction", "Success")
             const findUser = await userModel.findById({ _id: req.userId })
             if (!findUser) {
                 return sendResponse(res, HTTP_STATUS.NOT_FOUND, "Please sign in!");
@@ -243,6 +318,7 @@ class admin {
             return sendResponse(res, HTTP_STATUS.OK, "Successfully fetched data", view)
 
         } catch (error) {
+            log.createLogFile("/book/viewAllTransaction", "Failure")
             console.log(error);
             return sendResponse(res, HTTP_STATUS.INTERNAL_SERVER_ERROR, "Internal Server Error!");
         }
@@ -251,6 +327,10 @@ class admin {
         try {
             const { name, email, phone } = req.body;
             const query = {};
+            const validation = validationResult(req).array();
+            if (validation.length > 0) {
+                return sendResponse(res, HTTP_STATUS.UNPROCESSABLE_ENTITY, "Failed to add the book", validation);
+            }
             const findUser = await userModel.findById({ _id: req.userId });
             if (!findUser) {
                 return sendResponse(res, HTTP_STATUS.NOT_FOUND, "Please sign in!");
